% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastepath.R
\name{format_path}
\alias{format_path}
\title{Format text as a file path}
\usage{
format_path(text = character(0))
}
\arguments{
\item{text}{String. If not supplied this function reads the clipboard}
}
\value{
A well-formatted file path. When the file path is downstream of the
current working directory, a string representing a call to `file.path` with
a relative path is returned. Otherwise, the absolute file path is returned
}
\description{
Format text as a file path
}
