% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastepath.R
\name{to_path}
\alias{to_path}
\title{RStudio add-in function to convert a string to a well-formatted file path}
\usage{
to_path()
}
\description{
RStudio add-in function to convert a string to a well-formatted file path
}
